requirements.txt
textrazor
python-pptx
scikit-learn
speechrecognition
urllib3==1.26.7